better error handling on requests
